requirements.txt
et_xmlfile==2.0.0
openpyxl==3.1.5
python-slugify==8.0.4
RapidFuzz==3.10.1
text-unidecode==1.3
